feat(rootfs): support persistent running vsock-guest-exec
Run vsock-guest-exec based on systemd service in the guest, so that the host can send commands to be executed to the guest at any time.
The main function of this feature is: when the system is awakened, reset the guest's system time to the current host time.
In the future, there may be other functions, but this is beyond the scope of this project. You can follow at https://github.com/oomol-lab/ovm.

Signed-off-by: Black-Hole1 <bh@bugs.cc>