fix(kernel): files in overlayfs of container are missing
We found that when executing an overlay mount inside a Podman container, there is an issue: the middle layer contains a node_modules directory, but the upper layer cannot display this directory.

When we disable all overlay options, it returns to normal. Currently, there are no other side effects found from disabling the overlay options.

Signed-off-by: Kevin Cui <bh@bugs.cc>